pyproject.toml
[project]
name = "ode-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "diffrax>=0.7.2",
    "jax>=0.9.1",
    "matplotlib>=3.10.8",
    "numba>=0.61",
    "numba-cuda>=0.10",
    "numpy>=2.4.3",
    "scipy>=1.17.1",
    "triton>=3.6.0",
    "warp-lang>=1.10.0",
]

[project.optional-dependencies]
cuda12 = [
    "jax[cuda12]>=0.9.1",
]
cuda13 = [
    "jax[cuda13]>=0.9.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-benchmark>=5.2.3",
    "pytest-regressions>=2.10.0",
    "ruff>=0.15.5",
]

[tool.pytest.ini_options]
markers = ["slow: marks tests as slow (deselect with '-m \"not slow\"')"]
addopts = "-m 'not slow'"

[tool.ruff.lint]
# On top of the default `select`, enable imports sorting
extend-select = ["I"]

[tool.pyright]
typeCheckingMode = "recommended"
venv = ".venv"
venvPath = "."
